chore(native-base-components-image-slider): debugging jest runs
