+++ packages/native-base-image-slider/package.json
@@ -3,8 +3,9 @@
   "version": "0.0.0",
   "description": "",
   "main": "dist/index.js",
-  "types": "types/index.d.ts",
+  "types": "dist/index.d.ts",
   "scripts": {
+    "clean": "rimraf dist",
     "build": "tsc",
     "dev": "tsc --watch",
     "test": "jest",
@@ -53,6 +54,7 @@
     "react-native-svg-mock": "^2.0.0",
     "react-test-renderer": "^18.2.0",
     "react-testing-library": "^8.0.1",
+    "rimraf": "^5.0.1",
     "ts-jest": "^29.1.0",
     "ts-node": "^10.9.1",
     "typescript": "^4.9.4"
