--- packages/native-base-image-slider/types/ImageSlider/ImageSlider.d.ts
@@ -1,20 +1,0 @@
-import { FC } from 'react';
-
-interface ImageProps {
-  src: string;
-  caption?: string;
-}
-
-interface ImageSliderProps {
-  autoPlay?: boolean;
-  autoPlayInterval?: number;
-  data: ImageProps[];
-  separatorWidth?: number;
-  withPagination?: boolean;
-  width?: number;
-}
-
-declare const ImageSlider = FC<ImageSliderProps>;
-
-export { ImageSliderProps };
-export default ImageSlider;
