--- packages/native-base-image-slider/types/ImageSlider/index.d.ts
@@ -1,4 +1,0 @@
-import ImageSlider from './ImageSlider';
-
-export * from './ImageSlider';
-export default ImageSlider;
