--- packages/native-base-image-slider/types/index.d.ts
@@ -1,3 +1,0 @@
-import ImageSlider from './ImageSlider';
-
-export default ImageSlider;
