+++ packages/typography/package.json
@@ -5,6 +5,7 @@
   "main": "dist/index.js",
   "types": "types/index.d.ts",
   "scripts": {
+    "clean": "rimraf dist",
     "build": "tsc",
     "dev": "tsc --watch",
     "lint": "eslint .",
@@ -19,6 +20,7 @@
     "prettier": "^2.8.8",
     "react": "18.2.0",
     "react-native": "^0.71.8",
+    "rimraf": "^5.0.1",
     "typescript": "^4.9.4"
   },
   "peerDependencies": {
